requirements.txt
python-dotenv==1.0.1
Flask==3.0.2
Flask-SocketIO==5.3.6
openai==1.11.1

# easy gpt utils 相关依赖
# https://pypi.org/project/pinecone-client/
# https://github.com/pinecone-io/pinecone-python-client
pinecone-client==3.0.2